requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
decorator==5.1.1
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
graphdatascience==1.9
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langsmith==0.0.87
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.5
multimethod==1.11
mypy-extensions==1.0.0
neo4j==5.17.0
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==2.2.0
Pillow==9.5.0
plotly==5.15.0
protobuf==4.25.2
py==1.11.0
pyarrow==14.0.2
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
retry==0.9.2
rich==13.7.0
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
streamlit==1.23.1
streamlit-chat==0.0.2.2
streamlit-chat-media==0.0.4
tenacity==8.2.3
textdistance==4.6.1
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.2.0
validators==0.22.0
yarl==1.9.4
zipp==3.17.0
